gemini-cli/composio-cli-agent/requirements.txt
composio-core>=0.3.0
composio-openai>=0.3.0
typer[all]>=0.9.0
python-dotenv>=1.0.0
rich>=13.0.0
openai>=1.0.0
pytest>=7.0.0
google-search-results>=2.4.2